전체 수정 및 추가 -3
1. 회원가입 & 개인정보 수정시 닉네임 중복체크 구현
2. 개인정보 프로필 수정부분에서 닉네임만 수정할때를 위해 memberController  "/user/nick" 구현
3. 로그인과 로그아웃시 router 이동 실행 후 새로고침 되게
4. gitignore img & png 추가

+++ src/main/java/com/bookstore/bookstoreapi/member/MemberService.java
@@ -62,27 +62,31 @@ public class MemberService {
     @Transactional
     public MemberDTO editProfile(String nickName, MultipartFile multipartFile) {
         Member data = memberData();
-
-        String imgFileName = data.getMid() + "mid" + multipartFile.getOriginalFilename();
-        Path imgFilePath = Paths.get(uploadFolder+imgFileName);
-
         MemberDTO memberDTO = new MemberDTO();
 
-        if(multipartFile.getSize() !=0){
-            try{
-                if(data.getProfilePicture() != null){
-                    File file = new File(uploadFolder+data.getProfilePicture());
-                    file.delete();
+
+        if(multipartFile!=null){
+            String imgFileName = data.getMid() + "mid" + multipartFile.getOriginalFilename();
+            Path imgFilePath = Paths.get(uploadFolder+imgFileName);
+
+
+            if(multipartFile.getSize() !=0){
+                try{
+                    if(data.getProfilePicture() != null){
+                        File file = new File(uploadFolder+data.getProfilePicture());
+                        file.delete();
+                    }
+                    Files.write(imgFilePath, multipartFile.getBytes());
+                } catch (Exception e){
+                    e.printStackTrace();
                 }
-                Files.write(imgFilePath, multipartFile.getBytes());
-            } catch (Exception e){
-                e.printStackTrace();
+                data.updatePicture(imgFileName);
+                Member result = memberRepository.save(data);
+                memberDTO.setProfilePicture(result.getProfilePicture());
+                memberDTO.setNickName(result.getNickName());
             }
-            data.updatePicture(imgFileName);
-            Member result = memberRepository.save(data);
-            memberDTO.setProfilePicture(result.getProfilePicture());
-            memberDTO.setNickName(result.getNickName());
         }
+
 
         if(nickName!=null && nickName.length()>0){
             data.updateNickName(nickName);
